intermediate update of testing
